requirements.txt
aiohttp                  
aiosignal                
annotated-types         
antlr4-python3-runtime   
anyio                    
async-timeout            
attrs                    
backoff                  
beautifulsoup4           
certifi                  
cffi                     
chardet                  
charset-normalizer       
click                    
coloredlogs              
contourpy               
cryptography             
cycler                   
dataclasses-json         
Deprecated               
distro                   
dnspython                
effdet                   
emoji                    
exceptiongroup           
filelock                 
filetype                 
flatbuffers              
fonttools                
frozenlist               
fsspec                   
h11                      
httpcore                 
httpx                    
huggingface-hub          
humanfriendly            
idna                     
iopath                   
Jinja2                   
joblib                   
jsonpatch                
jsonpointer              
kiwisolver               
langchain               
langchain-core           
langdetect               
langsmith                
layoutparser             
loguru                   
lxml                     
MarkupSafe               
marshmallow             
matplotlib               
mpmath                   
multidict               
mypy-extensions          
networkx                
nltk                     
numpy                   
omegaconf                
onnx                     
onnxruntime              
openai                   
opencv-python            
packaging                
pandas                   
pdf2image                
pdfminer.six             
pdfplumber               
pikepdf                  
Pillow                   
pinecone-client         
pip                      
portalocker              
protobuf               
pycocotools              
pycparser                
pycryptodome             
pydantic                 
pydantic_core           
pyparsing                
pypdf                   
pypdfium2               
pytesseract              
python-dateutil          
python-iso639            
python-magic             
python-multipart         
pytz                     
PyYAML                   
rapidfuzz              
regex                    
requests                 
safetensors              
scipy                    
setuptools               
six                      
sniffio                  
soupsieve                
SQLAlchemy               
sympy                    
tabulate                
tenacity                 
tiktoken                 
timm                     
tokenizers               
torch                    
torchvision              
tqdm                     
transformers             
typing_extensions        
typing-inspect           
tzdata                   
unstructured            
unstructured-inference   
unstructured.pytesseract 
urllib3                  
wrapt                    
yarl                     